+++ src/teste.cpp
@@ -2,7 +2,7 @@
 #include <string>
 #include <vector>
 
-#include "funcionario.hpp"
+#include "funcionario.h"
 
 using namespace std;
 
@@ -12,7 +12,7 @@ int main() {
 	int i = 0;
 	int num;
 
-	Funcionario vet01("vet","Paulo Fernandes", "849872133420", "paulofernandes@hotmail.com","Rua florliz,554, Bairro das flores");
+	Funcionario vet01("0001", "Paulo",recepcionista, "849872133420", "paulofernandes@hotmail.com","Rua florliz,554, Bairro das flores");
 
 	cout << "----------Veterinario inserido----------" << endl;
 	cout << vet01.getCodigo() << " - " << vet01.getNome() << endl;
